factor out controls
