+++ frontend/src/pages/Main.tsx
@@ -1,47 +1,17 @@
 import { useDisclosure } from "@nextui-org/react";
 import React, { useEffect } from "react";
 import { Toaster } from "react-hot-toast";
-import CogTooth from "#/assets/cog-tooth";
 import ChatInterface from "#/components/chat/ChatInterface";
 import Errors from "#/components/Errors";
 import { Container, Orientation } from "#/components/Resizable";
 import Workspace from "#/components/Workspace";
 import LoadPreviousSessionModal from "#/components/modals/load-previous-session/LoadPreviousSessionModal";
 import SettingsModal from "#/components/modals/settings/SettingsModal";
-import AgentControlBar from "#/components/AgentControlBar";
-import AgentStatusBar from "#/components/AgentStatusBar";
-import VolumeIcon from "#/components/VolumeIcon";
+import Controls from "#/components/Controls";
 import Terminal from "#/components/terminal/Terminal";
 import Session from "#/services/session";
 import { getToken } from "#/services/auth";
 import { settingsAreUpToDate } from "#/services/settings";
-
-interface Props {
-  setSettingOpen: (isOpen: boolean) => void;
-}
-
-function Controls({ setSettingOpen }: Props): JSX.Element {
-  return (
-    <div className="flex w-full p-4 bg-neutral-900 items-center shrink-0 justify-between">
-      <div className="flex items-center gap-4">
-        <AgentControlBar />
-      </div>
-      <AgentStatusBar />
-
-      <div style={{ display: "flex", alignItems: "center" }}>
-        <div style={{ marginRight: "8px" }}>
-          <VolumeIcon />
-        </div>
-        <div
-          className="cursor-pointer hover:opacity-80 transition-all"
-          onClick={() => setSettingOpen(true)}
-        >
-          <CogTooth />
-        </div>
-      </div>
-    </div>
-  );
-}
 
 // React.StrictMode will cause double rendering, use this to prevent it
 let initOnce = false;
